test(server): split large handler test modules
Move the existing runtime-host and operator-monitor tests into scenario-specific include fragments while preserving their logical test paths and assertions.

Closes #1980
